Extract .meas directive building from GUI to simulation/measurement_builder (#639)
Move build_directive(), ANALYSIS_DOMAIN_MAP, and MEAS_TYPES from
GUI/meas_dialog.py to simulation/measurement_builder.py so that
measurement directive generation can be tested and reused without
importing PyQt6. GUI/meas_dialog.py re-exports for backward
compatibility.
